Refactor y mejora de componentes: eliminar DetailProvider, actualizar gestión de estado de login y optimizar avatar
