+++ src/mocks/users.json
@@ -5,28 +5,38 @@
         "password": "password123",
         "email": "will117@gmail.com",
         "role": "admin",
-        "firstName":"",
-        "lastName":""
+        "firstName":"Willy",
+        "lastName":"Carreño",
+        "date":"12/05/2024"
     },
     {
         "id": 2,
         "username": "michael3125",
         "password": "1",
         "email": "michael784@gmail.com",
-        "role": "user"
+        "role": "user",
+        "firstName":"Michael",
+        "lastName":"Zambrano",
+        "date":"15/06/2024" 
     },
     {
         "id": 3,
         "username": "ssamaeljs",
         "password": "ssamaeljs",
         "email": "ssamael.dev@unir.net",
-        "role": "admin"
+        "role": "admin",
+        "firstName":"Samuel",
+        "lastName":"Lopez",
+        "date":"20/07/2024"
     },
     {
         "id": 4,
         "username": "cdcp",
         "password": "abc123",
         "email": "cdcp@unir.net",
-        "role": "admin"
+        "role": "admin",
+        "firstName":"Cristian",
+        "lastName":"Chiguano",
+        "date":"25/08/2025"
     }
 ]
