--- src/shared/context/DetailContext/DetailsContext.jsx
@@ -1,3 +1,0 @@
-import { createContext } from "react";
-
-export const DetailContext = createContext();
