+++ src/shared/hooks/useAvatar.js
